+++ threedi_models_and_simulations/widgets/simulation_results.py
@@ -136,6 +136,7 @@ class SimulationResults(uicls, basecls):
             simulation = tc.fetch_simulation(sim_id)
             simulation_name = simulation.name.replace(" ", "_")
             simulation_model_id = int(simulation.threedimodel_id)
+            results_dir, gridadmin_downloads, gridadmin_downloads_gpkg = None, None, None
             try:
                 model_3di = tc.fetch_3di_model(simulation_model_id)
                 gridadmin_downloads = tc.fetch_3di_model_gridadmin_download(simulation_model_id)
@@ -167,7 +168,9 @@ class SimulationResults(uicls, basecls):
                     if not results_dir:
                         self.plugin_dock.communication.show_warn(warn_msg)
                         return
+                gridadmin_downloads_gpkg = tc.fetch_3di_model_geopackage_download(simulation_model_id)
             except ApiException as e:
+                error_msg = extract_error_message(e)
                 if e.status == 404:
                     warn_msg = (
                         "The 3Di model to which these results belong is owned by an organisation for which "
@@ -179,15 +182,18 @@ class SimulationResults(uicls, basecls):
                     )
                     self.plugin_dock.communication.show_warn(warn_msg)
                     results_dir = self.pick_results_destination_dir()
-                    if not results_dir:
-                        return
-                    gridadmin_downloads = None
+                elif "Geopackage file not found" in error_msg:
+                    pass
                 else:
                     raise e
+            if not results_dir:
+                return
             simulation_subdirectory = os.path.join(results_dir, f"sim_{sim_id}_{simulation_name}")
             downloads = tc.fetch_simulation_downloads(sim_id)
             if gridadmin_downloads is not None:
                 downloads.append(gridadmin_downloads)
+            if gridadmin_downloads_gpkg is not None:
+                downloads.append(gridadmin_downloads_gpkg)
             downloads.sort(key=lambda x: x[-1].size)
             self.last_progress_item = self.tv_model.item(current_row, self.PROGRESS_COLUMN_IDX)
         except ApiException as e: